Perform maven package release as part of npx semantic-release
The main benefit of this is that we will now only try publishing to
maven (and npm) repositories if there has been a change in the version
number - which is decided by the semantic-release tool.

It will also fail the release if we successfully publish to npm but fail
to publish to maven, thus preventing the automatic PRs from being
created in either the core or b7 repos - so there's no chance of them
getting out of sync.